Relocate identifier-case utilities into plugin tree
