+++ src/plugin/tests/identifier-case-utils.test.js
@@ -6,7 +6,7 @@ import {
     formatIdentifierCase,
     isIdentifierCase,
     normalizeIdentifierCase
-} from "../../shared/identifier-case/identifier-case-utils.js";
+} from "../src/identifier-case/identifier-case-utils.js";
 
 test("normalisation preserves prefixes and numeric suffixes", () => {
     const normalized = normalizeIdentifierCase("global.hp_max_2");
